+++ packages/flutter/test/components/area_chart/pl_area_chart_test.dart
@@ -87,7 +87,7 @@ void main() {
 
       expect(
         tester.getSemantics(find.bySemanticsLabel('Chart')).value,
-        r'New $4000, Renewed $16000',
+        r'New: $4000. Renewed: $16000',
       );
     });